+++ Assets/GameLogic/Spells/Modificators/QuickModificator.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class QuickModificator:  AbstractSpellModificator {
-	public double speedFactor = 1.0;
-	public double weakFactor = 1.0;
+	public float speedFactor = 1.0f;
+	public float weakFactor = 1.0f;
 	public override string Name {
 		get {
 			return "quick";
